Update dataTables.responsive.js
As mentioned in your code the calculation of the minWidth per cell is not 100% perfect. I ran in some overflow-issues with large tables ( 30+ cols). To keep the good performance and to fix this i added 20px per column. It works perfect in my case. This might be a quirky work-around, but the only bad result might be a greater cell-padding. 

This fixes https://github.com/DataTables/Responsive/issues/20