Modify _columnVisibility to use "pri-xxx" to set column priority
Adding the class "pri-xxx" to a column will set the priority of that column.  All columns marked as "all" will still be shown first, then instead of displaying the remaining columns based on left to right sort order, it will show them based on their priority.  For example, a column with a class of "pri-1" will display before a column with a class of "pri-2", even if the column with class "pri-2" comes before the other column.