FIX: nested tables break responsivnes of other columns
When there is a nested table in a cell and it's loaded using ajax the
number of all found cells in a row is higher than number of columns in a
table. That breaks hiding cells at the end of a row.
I'm taking only immediate children or TR not all